requirements.txt
elasticsearch
python-nmap
requests
urllib3
humanfriendly
pysmb